backend/pyproject.toml
[project]
name = "a-share-strategy-assistant"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.0",
  "alembic>=1.13.0",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.4.0",
  "redis>=5.0.0",
  "celery>=5.4.0",
  "akshare>=1.15.0",
  "httpx>=0.27.0",
  "pyarrow>=15.0.0",
  "pandas>=2.0.0",
  "h5py>=3.10.0",
  "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "ruff>=0.6.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
